requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-amazon==8.4.0
pymongo==4.6.1
pika==1.3.2